@@ -1,8 +1,8 @@
 ext {
-    compileSdkVersion = 26
-    buildToolsVersion = '26.0.2'
-    targetSdkVersion = 22
-    minSdkVersion = 14
+    compileSdkVersion = 28
+    buildToolsVersion = '28.0.3'
+    targetSdkVersion = 28
+    minSdkVersion = 21
 
     sourceCompatibilityVersion = JavaVersion.VERSION_1_8
     targetCompatibilityVersion = JavaVersion.VERSION_1_8
@@ -12,39 +12,29 @@ ext {
 
     dependency = [
             // Gradle plugins
-            spoonGradlePlugin          : 'com.stanfy.spoon:spoon-gradle-plugin:1.0.4',
-            androidGradlePlugin        : 'com.android.tools.build:gradle:3.0.0',
-
-            // Mopub
-            mopub                      : 'com.mopub:mopub-sdk-native-static:4.11.0@aar',
+            androidGradlePlugin        : 'com.android.tools.build:gradle:3.6.0',
 
             // Sample App
             leakCanary                 : 'com.squareup.leakcanary:leakcanary-android:1.5.3',
             leakCanaryNoOp             : 'com.squareup.leakcanary:leakcanary-android-no-op:1.5.3',
-            appCompat                  : 'com.android.support:appcompat-v7:26.1.0',
+            appCompat                  : 'androidx.appcompat:appcompat:1.2.0',
 
             // Tests
-            spoon                      : 'com.squareup.spoon:spoon-client:1.3.2',
-            espresso                   : 'com.android.support.test.espresso:espresso-core:3.0.1',
-            androidTestingSupport      : 'com.android.support.test:runner:1.0.1',
-            junit                      : 'junit:junit:4.12',
-            equalsVerifier             : 'nl.jqno.equalsverifier:equalsverifier:2.2.1',
+            androidTestingSupport      : 'androidx.test.ext:junit:1.1.2',
+            junit                      : 'junit:junit:4.13.2',
+            equalsVerifier             : 'nl.jqno.equalsverifier:equalsverifier:2.5.2',
             robolectric                : 'org.robolectric:robolectric:3.2.2',
-            mockito                    : 'org.mockito:mockito-core:2.2.0',
-            dexmakerMockito            : 'com.linkedin.dexmaker:dexmaker-mockito:2.2.0',
+            mockito                    : 'org.mockito:mockito-core:2.28.1',
 
             // Other
-            picasso                    : 'com.squareup.picasso:picasso:2.5.2',
-            supportCompatV4            : 'com.android.support:support-compat:26.1.0',
-            supportCoreV4              : 'com.android.support:support-core-ui:26.1.0',
+            picasso                    : 'com.squareup.picasso:picasso:2.8',
             retrofit                   : 'com.squareup.retrofit2:retrofit:2.3.0',
             retrofitMock               : 'com.squareup.retrofit2:retrofit-mock:2.3.0',
             retrofitGsonConverter      : 'com.squareup.retrofit2:converter-gson:2.3.0',
             gson                       : 'com.google.code.gson:gson:2.7',
-            okHttp                     : 'com.squareup.okhttp3:okhttp:3.8.0',
-            okHttpInterceptor          : 'com.squareup.okhttp3:logging-interceptor:3.8.0',
-            design                     : 'com.android.support:design:26.1.0',
-            tweetText                  : 'com.twitter:twitter-text:1.14.7',
-            recyclerview               : 'com.android.support:recyclerview-v7:26.1.0'
+            okHttp                     : 'com.squareup.okhttp3:okhttp:3.8.1',
+            okHttpInterceptor          : 'com.squareup.okhttp3:logging-interceptor:3.8.1',
+            design                     : 'com.google.android.material:material:1.3.0',
+            tweetText                  : 'com.twitter:twitter-text:1.14.7'
     ]
 }
